requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
google-generativeai==0.4.1
pydantic==2.6.3
httpx==0.27.0
crawl4ai==0.3.5
pandas==2.2.1
maturin==1.5.0
streamlit==1.32.0